Computer Vision
Consider a rotation about the X axis of 1.1 radians, followed by a rotation about the Y axis of
-0.5 radians, followed by a rotation about the Z axis by 0.1 radians (this order of rotations is
called the “XYZ fixed angles” convention).
a. Give the 3x3 rotation matrix corresponding to the rotations above.
b. Since the rotation matrix is “orthonormal” matrix (i.e., a square matrix whose rows and
columns are orthogonal unit vectors), its inverse is equal to its transpose. Show this.
c. Give the 3x3 rotation matrix where the same rotations described in part (a) are done in
the opposite order; i.e., first a rotation about the Z axis of by 0.1 radians, followed by a
rotation about the Y axis of -0.5 radians, followed by a rotation about the X axis by 1.1
radians (this convention is called “ZYX fixed angles”). The matrix should be different.
2. A camera observes the following 7 points, defined in WORLD coordinates (meters):
 6.8158 7.8493 9.9579 8.8219 9.5890 10.8082 13.2690
 -35.1954 -36.1723 -25.2799 -38.3767 -28.8402 -48.8146 -58.0988
 43.0640 43.7815 40.1151 46.6153 42.2858 56.1475 59.1422
The pose of the camera with respect to the world is given by the following:
• Translation of camera origin with respect to the world is (10,-25,40) in meters.
• Orientation of the camera with respect to the world is given by the angles provided in
problem #1.
a) Compute the homogeneous transformation matrix that represents the pose of the
camera with respect to the world, HW
C , assuming that the convention being used is
“XYZ fixed angles”.
b) Compute the homogeneous transformation matrix that represents the pose of the
world with respect to the camera, HC
W .
c) Assume that the size of the image is 256 columns (width) by 170 rows (height), with
the optical center at the image center. The effective focal length is 400 pixels. Write
the intrinsic camera calibration matrix K.
d) Create a blank (zeros) image, and project the 7 points to the image. Write white dots
into the image at those points.
e) Using Matlab’s “line” function, draw lines between the points on the image.
Namely draw a line from point 1 to point 2, another line from point 2 to point 3, and
so forth. Show the resulting image